+++ src/lang/hooks.ml
@@ -14,7 +14,7 @@ let liq_libs_dir = ref (fun () -> raise Not_found)
 let log_path = ref None
 let collect_after = ref (fun fn -> fn ())
 
-module type Regexp_t = Regexp.T
+module type Regexp_t = Regexp.T with type t := Regexp.t
 
 let regexp = Regexp.regexp_ref