+++ src/lang/lang.ml
@@ -1,2 +1,3 @@
 include Lang_core
 include Lang_error
+include Lang_regexp