+++ src/lang/regexp.mli
@@ -1,15 +1,15 @@
-type sub = ..
 type t = ..
 type flag = [ `i | `g | `s | `m ]
+type sub = { matches : string option list; groups : (string * string) list }
 
 module type T = sig
+  type t
+
   val regexp : ?flags:flag list -> string -> t
   val regexp_or : ?flags:flag list -> string list -> t
   val split : ?pat:string -> ?rex:t -> string -> string list
   val exec : ?pat:string -> ?rex:t -> string -> sub
   val test : ?pat:string -> ?rex:t -> string -> bool
-  val num_of_subs : sub -> int
-  val get_substring : sub -> int -> string
 
   val substitute :
     ?pat:string -> ?rex:t -> subst:(string -> string) -> string -> string
@@ -18,6 +18,8 @@ module type T = sig
     ?pat:string -> ?rex:t -> subst:(string -> string) -> string -> string
 end
 
-include T
+module type Regexp_t = T with type t := t
 
-val regexp_ref : (module T) ref
+val regexp_ref : (module Regexp_t) ref
+
+include T with type t := t