+++ CHANGES.md
@@ -22,6 +22,8 @@ Changed:
 - Added support for a Javascript build an interpreter.
 - Removed support for `%define` variables, superseeded by support for actual
   variables in encoders.
+- Reimplemented `harbor` http handler API to be more flexible. Added a new
+  node/express-like registration and middleware API (#2599).
 - Switched default persistence for cross and fade-related overrides
   to follow documented behavior. By default, `"liq_fade_out"`, `"liq_fade_skip"`,
   `"liq_fade_in"`, `"liq_cross_duration"` and `"liq_fade_type"` now all reset on